requirements.txt
fastapi
pydantic
uvicorn
xgboost
scikit-learn
requests
sqlalchemy 
pandas 
psycopg2-binary 
python-dotenv
jupyter